+++ src/api/objects/tag.ts
@@ -4,6 +4,10 @@
  */
 import { mastodon } from "masto";
 
+import { countValues } from "../../helpers/collection_helpers";
+import { MastoApi } from "../../api/api";
+import { StringNumberDict } from "../../types";
+
 const BROKEN_TAG = "<<BROKEN_TAG>>";
 
 
@@ -18,4 +22,12 @@ export function repairTag(tag: mastodon.v1.Tag): mastodon.v1.Tag {
 
     tag.url = tag.url.toLowerCase();
     return tag;
+};
+
+
+// Count how many times th euser has used a given tag
+export async function participatedHashtags(): Promise<StringNumberDict> {
+    const recentToots = await MastoApi.instance.getUserRecentToots();
+    const hashtags = recentToots.flatMap(toot => toot.realToot().tags || []);
+    return countValues<mastodon.v1.Tag>(hashtags, (tag) => tag.name);
 };