+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* Add HashtagParticipationScorer and import recent toots that contain hashtags the user participates in often in the feed
 * Poll for old records after initial load with `setInterval()`. Add `Config.backgroundLoadIntervalMS` and `Config.maxRecordsForFeatureScoring` variables
 
 ### v0.22.4
@@ -78,7 +79,7 @@
 * Remove `TrendingLinksScorer.trendingLinks` property
 * Stop mutating the `Toot` objects in `FollowedTagsScorer` (now happens in `Toot.setDependentProperties()`)
 * Set `Toot.muted` where necessary in `Toot.setDependentProperties()`
-* Add Toot.setDependentProps() static method to set props for array of Toots. move the setting of those props to the calls like MastoApi.fetchHomeFeed() that do the initial Toot object building
+* Add `Toot.setDependentProps()` static method to set props for array of `Toot` objectss. move the setting of those props to the calls like `MastoApi.fetchHomeFeed()` that do the initial Toot object building
 
 ### v0.20.7
 * Consolidate all `Toot` post processing to `Toot.setDependentProperties()` method